aashutosh/aashutosh_jewelerys_project/requirements.txt
# === CORE DEPENDENCIES ===
Django==5.0.2
Pillow==10.2.0
python-decouple==3.8

# === DATABASE SUPPORT ===
psycopg2-binary==2.9.9          # PostgreSQL connector
mysqlclient==2.2.0              # MySQL connector
dj-database-url==2.1.0          # Database URL parsing

# === STATIC FILES & MEDIA ===
whitenoise==6.6.0               # Static file serving
django-storages==1.14.2         # Cloud storage support
boto3==1.34.0                   # AWS S3 storage

# === SECURITY & AUTHENTICATION ===
django-cors-headers==4.3.1      # CORS headers
bcrypt==4.1.2                   # Password hashing
python-decouple==3.8            # Environment variables

# === DEVELOPMENT TOOLS ===
django-debug-toolbar==4.3.0     # Debugging toolbar
django-extensions==3.2.3        # Extended management commands
ipython==8.22.0                 # Enhanced Python shell
python-dotenv==1.0.0            # .env file support

# === PAYMENT PROCESSING ===
stripe==7.7.0                   # Stripe payments
razorpay==1.4.3                 # Razorpay payments

# === EMAIL & NOTIFICATIONS ===
django-sendgrid-v5==0.9.2       # SendGrid email
celery==5.3.4                   # Background tasks
redis==5.0.1                    # Redis for Celery

# === SEARCH & FILTERING ===
django-filter==23.5             # Advanced filtering
django-haystack==3.2.1          # Search functionality
whoosh==2.7.4                   # Search engine

# === FORM & VALIDATION ===
crispy-forms==2.1               # Better form rendering
crispy-bootstrap5==0.7           # Bootstrap 5 forms
django-phonenumber-field==7.1.0 # Phone number validation

# === API SUPPORT ===
djangorestframework==3.14.0     # REST API framework
django-cors-headers==4.3.1      # CORS for APIs
markdown==3.5.2                 # Markdown support

# === DEPLOYMENT & SERVERS ===
gunicorn==21.2.0                # Production WSGI server
uvicorn==0.24.0                 # ASGI server
daphne==4.0.0                   # Django Channels server

# === MONITORING & LOGGING ===
sentry-sdk==1.40.0              # Error tracking
django-axes==6.0.0              # Security monitoring

# === E-COMMERCE SPECIFIC ===
django-cart==2.2.0              # Shopping cart functionality
django-wishlist==0.3.0          # Wishlist feature
django-countries==7.5.1         # Country field

# === UTILITIES ===
requests==2.31.0                # HTTP requests
beautifulsoup4==4.12.2          # HTML parsing
pandas==2.1.4                   # Data analysis
openpyxl==3.1.2                 # Excel support

# === TESTING ===
pytest==7.4.3                   # Testing framework
pytest-django==4.7.0            # Django pytest integration
factory-boy==3.3.0              # Test data factory
coverage==7.3.2                 # Test coverage